+++ src/models/shared/links.rs
